+++ tests/org.corpus_tools.hexatomic.it.tests/src/main/java/org/corpus_tools/hexatomic/it/tests/TestGridEditor.java
@@ -1731,6 +1731,48 @@ public class TestGridEditor {
     dialog.close();
   }
 
+  /**
+   * Regression test for https://github.com/hexatomic/hexatomic/issues/259.
+   * 
+   * <p>
+   * Tests that the annotation name, not the display name, is shown on columns that are secondary
+   * columns (e.g., "namespace::name (2)").
+   * </p>
+   * 
+   * @throws InterruptedException If the thread extracting the label text in the given dialog is
+   *         interrupted, or if there is an error during execution.
+   * @throws ExecutionException See InterruptedException
+   */
+  @Test
+  void testShowCorrectAnnotationNameInDialog() throws InterruptedException, ExecutionException {
+    openOverlapExample();
+
+    SWTNatTableBot tableBot = new SWTNatTableBot();
+    SWTBotNatTable table = tableBot.nattable();
+
+    assertEquals("five::span_1 (2)", table.getCellDataValueByPosition(0, 4));
+
+    // Test cell menu behaviour
+    table.click(1, 4);
+    table.contextMenu(1, 4).contextMenu(GridEditor.CHANGE_ANNOTATION_NAME_POPUP_MENU_LABEL).click();
+
+    bot.waitUntil(Conditions.shellIsActive(RENAME_DIALOG_TITLE));
+    SWTBotShell dialog = tableBot.shell(RENAME_DIALOG_TITLE);
+    assertNotNull(dialog);
+    assertDialogTexts(dialog, "five::span_1");
+    dialog.close();
+
+    // Test column menu
+    table.click(0, 4);
+    table.contextMenu(0, 4).contextMenu(GridEditor.CHANGE_ANNOTATION_NAME_POPUP_MENU_LABEL).click();
+
+    bot.waitUntil(Conditions.shellIsActive(RENAME_DIALOG_TITLE));
+    dialog = tableBot.shell(RENAME_DIALOG_TITLE);
+    assertNotNull(dialog);
+    assertDialogTexts(dialog, "five::span_1");
+    dialog.close();
+  }
+
   private void assertDialogTexts(SWTBotShell dialog, String qualifiedName)
       throws InterruptedException, ExecutionException {
     String namespace = null;