+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Optimize column lifecycle in the Grid Editor, so that emptied columns don't disappear (#258, #346)
+- Let annotation rename dialog display the actual annotation name, not the display name of the column (#259)
 
 ## [0.8.0] - 2021-11-16